Refactor to move prio code into a prio.py file
